+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Remove files from Edit Mods Dialog that don't exist anymore
 - Add Show In File Explorer button to Edit Mods right click menu
+- Allow importing packs without project or file id in manifest