+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Remove files from Edit Mods Dialog that don't exist anymore
 - Add Show In File Explorer button to Edit Mods right click menu
 - Allow importing packs without project or file id in manifest
+- Allow changing sort order when adding Curse mods